requirements.txt
fastapi
uvicorn
pydantic
pandas
sqlalchemy
pyodbc
urllib3 == 2.2.3
rapidfuzz
email_validator  == 2.2.0
psycopg2-binary